ZQS-257 Move split_circuit to zquantum.core.circuits
