+++ src/python/zquantum/core/circuits/__init__.py
@@ -129,7 +129,7 @@ from ._builtin_gates import (
     Z,
     builtin_gate_by_name,
 )
-from ._circuit import Circuit
+from ._circuit import Circuit, split_circuit
 from ._compatibility import new_circuit_from_old_circuit
 from ._gates import (
     ControlledGate,
